requirements.txt
# Core Dependencies
neo4j==5.14.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LangChain Integration
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10

# LLM Providers
google-generativeai>=0.8.0

# Utilities
python-dateutil==2.8.2
httpx==0.25.2
tenacity==8.2.3

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
ruff==0.1.8

